Fix circular dependency by moving DefaultTemplateLiteralProxy to main package
- Move DefaultTemplateLiteralProxy from templatelit to main package
- Removes circular dependency between packages
- All tests pass, decycle check passes
