+++ rhino/src/main/java/org/mozilla/javascript/DefaultTemplateLiteralProxy.java
@@ -4,13 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.javascript.templatelit;
-
-import org.mozilla.javascript.Context;
-import org.mozilla.javascript.Scriptable;
-import org.mozilla.javascript.ScriptableObject;
-import org.mozilla.javascript.TemplateLiteralProxy;
-import org.mozilla.javascript.Undefined;
+package org.mozilla.javascript;
 
 /**
  * Default implementation of template literal proxy. This implementation contains the template
